scripts/3d_pipeline/requirements.txt
torch>=2.1.0
diffusers>=0.27.0
huggingface-hub
trimesh[easy]
Pillow>=10.0.0
pygltflib
supabase>=2.4.0
python-dotenv
rembg>=2.0.50
usd-core>=24.3
